refactor: migrate the auth applications to the user module
