+++ bandlog-shared/src/main/java/net/effize/bandlog/shared/auth/AuthUser.java
@@ -1,0 +1,10 @@
+package net.effize.bandlog.shared.auth;
+
+
+public record AuthUser(
+        Long id,
+        String supabaseUserId,
+        String email,
+        String nickname
+) {
+}
