+++ bandlog-shared/src/main/java/net/effize/bandlog/shared/auth/SupabaseAuthenticationPrincipal.java
@@ -1,0 +1,5 @@
+package net.effize.bandlog.shared.auth;
+
+
+public record SupabaseAuthenticationPrincipal(String supabaseUserId, String email) {
+}
