+++ bandlog-user/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
 dependencies {
     implementation(project(":bandlog-shared"))
     implementation("org.springframework.boot:spring-boot-starter-web")
+    implementation("org.springframework.boot:spring-boot-starter-security")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     testImplementation("org.springframework.boot:spring-boot-starter-test")
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
